hilder_ali_crawler/requirements.txt
aiohttp==3.4.4
async-timeout==3.0.0
asyncio==3.4.3
attrs==18.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
gevent==1.3.6
greenlet==0.4.15
idna==2.7
idna-ssl==1.1.0
multidict==4.4.2
pika==0.12.0
pycparser==2.19
pymongo==3.7.1
PyYAML==3.13
requests==2.19.1
urllib3==1.23
yarl==1.2.6
